Update formalisms for the posited encoding approach
The syntax and palette expansion rules have been updated, partially to reflect
the current state of the mechanization, but also including aspects that haven't
been fleshed out yet in agda but which accurately reflect the current approach
of having the encoding be posited. The rules should now be accurate to use as
a guide for further mechanization and discussion.